+++ packages/browser/src/client.ts
@@ -1,37 +1,11 @@
-import { API, BaseClient, Scope } from '@sentry/core';
-import { DsnLike, Event, EventHint } from '@sentry/types';
+import { BaseClient, Scope } from '@sentry/core';
+import { Event, EventHint } from '@sentry/types';
 import { getGlobalObject, logger } from '@sentry/utils';
 
 import { BrowserBackend, BrowserOptions } from './backend';
+import { injectReportDialog, ReportDialogOptions } from './helpers';
 import { Breadcrumbs } from './integrations';
 import { SDK_NAME, SDK_VERSION } from './version';
-
-/**
- * All properties the report dialog supports
- */
-export interface ReportDialogOptions {
-  [key: string]: any;
-  eventId?: string;
-  dsn?: DsnLike;
-  user?: {
-    email?: string;
-    name?: string;
-  };
-  lang?: string;
-  title?: string;
-  subtitle?: string;
-  subtitle2?: string;
-  labelName?: string;
-  labelEmail?: string;
-  labelComments?: string;
-  labelClose?: string;
-  labelSubmit?: string;
-  errorGeneric?: string;
-  errorFormEntry?: string;
-  successMessage?: string;
-  /** Callback after reportDialog showed up */
-  onLoad?(): void;
-}
 
 /**
  * The Sentry Browser SDK Client.
@@ -110,14 +84,6 @@ export class BrowserClient extends BaseClient<BrowserBackend, BrowserOptions> {
       return;
     }
 
-    const script = document.createElement('script');
-    script.async = true;
-    script.src = new API(dsn).getReportDialogEndpoint(options);
-
-    if (options.onLoad) {
-      script.onload = options.onLoad;
-    }
-
-    (document.head || document.body).appendChild(script);
+    injectReportDialog(options);
   }
 }
