+++ packages/browser/src/index.ts
@@ -3,8 +3,8 @@ export * from './exports';
 import { Integrations as CoreIntegrations } from '@sentry/core';
 import { getGlobalObject } from '@sentry/utils';
 
-import { Breadcrumbs, GlobalHandlers, LinkedErrors, TryCatch, UserAgent } from './integrations';
-import { BaseTransport, FetchTransport, XHRTransport } from './transports';
+import * as BrowserIntegrations from './integrations';
+import * as Transports from './transports';
 
 let windowIntegrations = {};
 
@@ -19,17 +19,7 @@ if (_window.Sentry && _window.Sentry.Integrations) {
 const INTEGRATIONS = {
   ...windowIntegrations,
   ...CoreIntegrations,
-  Breadcrumbs,
-  GlobalHandlers,
-  LinkedErrors,
-  TryCatch,
-  UserAgent,
+  ...BrowserIntegrations,
 };
 
-const TRANSPORTS = {
-  BaseTransport,
-  FetchTransport,
-  XHRTransport,
-};
-
-export { INTEGRATIONS as Integrations, TRANSPORTS as Transports };
+export { INTEGRATIONS as Integrations, Transports };
